custom-elements/custom-element-lifecycle/types-of-callbacks/left-view-callback.html

<!DOCTYPE html>
<html>
<head>
<title>leftView callback of a custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="leftView callback must be enqueued whenever custom element is removed from the document and this document has a browsing context">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.leftViewCallback = function () {
        callbackCalled = true;
    }
    
    assert_false(callbackCalled, 'Callback leftView was called after prototype creation '
            + 'but should not');
    
    var XFoo = doc.register('x-foo-1', {prototype: proto});
    
    assert_false(callbackCalled, 'Callback leftView was called but should not');
    
    doc.body.innerHTML = '<x-foo-1 id="xfoo1"></x-foo-1>';

    assert_false(callbackCalled, 'Callback leftView was called');
    
    var xfoo1 = doc.querySelector('#xfoo1');
    
    doc.body.removeChild(xfoo1);
    
    assert_false(callbackCalled, 'Callback leftView was called for the document '
            + 'with no browsing context');
    
}, 'Test leftView callback when custom element is instantiating in HTML. '
    + 'Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.leftViewCallback = function () {
        callbackCalled = true;
    }
    
    assert_false(callbackCalled, 'Callback leftView was called after prototype '
            + 'creation but should not');
        
    doc.body.innerHTML = '<x-foo-2 id="xfoo2"></x-foo-2>';
    
    assert_false(callbackCalled, 'Callback leftView was called but should not');
    
    var XFoo = doc.register('x-foo-2', {prototype: proto});

    assert_false(callbackCalled, 'Callback leftView was called');
    
    var xfoo2 = doc.querySelector('#xfoo2');
    
    doc.body.removeChild(xfoo2);
    
    assert_false(callbackCalled, 'Callback leftView was called for the document '
            + 'with no browsing context');
    
}, 'Test leftView callback when custom element is instantiating in HTML before '
    + 'registration of a custom element. Document has no browsing context');
    

testInIFrame('../../resources/x-foo.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.leftViewCallback = function () {
        callbackCalled = true;
    }
    
    assert_false(callbackCalled, 'Callback leftView was called after prototype '
            + 'creation but should not');
            
    var XFoo = doc.register('x-foo', {prototype: proto});

    assert_false(callbackCalled, 'Callback leftView was called');
    
    var xfoo = doc.querySelector('#xfoo');
    
    doc.body.removeChild(xfoo);
    
    assert_true(callbackCalled, 'Callback leftView was not called for the document '
            + 'with browsing context');

}, 'Test leftView callback. Document has browsing context');

    
</script>
</body>
</html>
